+++ PLC_esp8266/main/LogicProgram/Inputs/TimerMSecs.cpp
@@ -5,7 +5,6 @@
 #include "LogicProgram/Serializer/Record.h"
 #include "esp_err.h"
 #include "esp_log.h"
-#include "esp_timer.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -31,7 +30,7 @@ void TimerMSecs::SetTime(uint32_t delay_time_ms) {
         delay_time_ms = TimerMSecs::max_delay_time_ms;
     }
     this->delay_time_us = delay_time_ms * 1000LL;
-    str_size = sprintf(this->str_time, "%u", delay_time_ms);
+    str_size = sprintf(this->str_time, "%u", (unsigned int)delay_time_ms);
 
     ESP_LOGD(TAG_TimerMSecs, "ctor, str_time:%s", this->str_time);
 }
